Add a new ReconnectionStrategy which always reconnects after X time, e.g. after 10 seconds.

+++ xmpp-core-client/src/main/java/rocks/xmpp/core/session/TruncatedBinaryExponentialBackoffStrategy.java
@@ -64,7 +64,7 @@ final class TruncatedBinaryExponentialBackoffStrategy implements ReconnectionStr
     }
 
     @Override
-    public int getNextReconnectionAttempt(int attempt) {
+    public long getNextReconnectionAttempt(int attempt) {
         // For the first attempt choose a random number between 0 and 60.
         // For the second attempt choose a random number between 0 and 180.
         // For the third attempt choose a random number between 0 and 420.
